+++ collections/maps/ordered.go
@@ -12,13 +12,13 @@ type Ordered[K cmp.Ordered, V any] struct {
 
 // NewOrdered returns an [Ordered] containing the given key/value pairs.
 func NewOrdered[K cmp.Ordered, V any](pairs ...Pair[K, V]) *Ordered[K, V] {
-	return newOrderedFromPairs[K, V, *Ordered[K, V]](pairs)
+	return orderedFromPairs[K, V, *Ordered[K, V]](pairs)
 }
 
-// NewOrderedFromIter returns an [Ordered] containing the key/value pairs from
-// the given iterator.
-func NewOrderedFromIter[K cmp.Ordered, V any](pairs iter.Seq2[K, V]) *Ordered[K, V] {
-	return newOrderedFromIter[K, V, *Ordered[K, V]](pairs)
+// NewOrderedFromSeq returns an [Ordered] containing the key/value pairs yielded
+// by the given sequence.
+func NewOrderedFromSeq[K cmp.Ordered, V any](seq iter.Seq2[K, V]) *Ordered[K, V] {
+	return orderedFromSeq[K, V, *Ordered[K, V]](seq)
 }
 
 // Set sets the value associated with the given key.
@@ -83,34 +83,34 @@ func (m *Ordered[K, V]) Project(transform func(K, V) (K, V, bool)) *Ordered[K, V
 	return orderedProject(m, transform)
 }
 
-// All returns an iterator that yields all key/value pairs in the map in order.
+// All returns a sequence that yields all key/value pairs in the map in order.
 func (m *Ordered[K, V]) All() iter.Seq2[K, V] {
 	return orderedAll[K, V](m)
 }
 
-// Keys returns an iterator that yields all keys in the map in order.
+// Keys returns a sequence that yields all keys in the map in order.
 func (m *Ordered[K, V]) Keys() iter.Seq[K] {
 	return orderedKeys[K, V](m)
 }
 
-// Values returns an iterator that yields all values in the map in order.
+// Values returns a sequence that yields all values in the map in order.
 func (m *Ordered[K, V]) Values() iter.Seq[V] {
 	return orderedValues[K, V](m)
 }
 
-// Reverse returns an iterator that yields all key/value pairs in the map in
+// Reverse returns a sequence that yields all key/value pairs in the map in
 // reverse order.
 func (m *Ordered[K, V]) Reverse() iter.Seq2[K, V] {
 	return orderedReverse[K, V](m)
 }
 
-// ReverseKeys returns an iterator that yields all keys in the map in reverse
+// ReverseKeys returns a sequence that yields all keys in the map in reverse
 // order.
 func (m *Ordered[K, V]) ReverseKeys() iter.Seq[K] {
 	return orderedReverseKeys[K, V](m)
 }
 
-// ReverseValues returns an iterator that yields all values in the map in reverse
+// ReverseValues returns a sequence that yields all values in the map in reverse
 // order.
 func (m *Ordered[K, V]) ReverseValues() iter.Seq[V] {
 	return orderedReverseValues[K, V](m)