+++ collections/maps/orderedbykey_test.go
@@ -18,7 +18,7 @@ func TestOrderedByKey(t *testing.T) {
 	testOrderedMap(
 		t,
 		NewOrderedByKey[reverseOrderedString, int],
-		NewOrderedByKeyFromIter[reverseOrderedString, int],
+		NewOrderedByKeyFromSeq[reverseOrderedString, int],
 		reverseOrderedString.Compare,
 		rapid.Custom(
 			func(t *rapid.T) reverseOrderedString {